+++ changelog.md
@@ -4,13 +4,14 @@ This project does NOT (yet) adhere to [Semantic Versioning](http://semver.org/).
 It is running as an Alfa/Beta, once v1 is reached we will start supporting semver.
 For now, please use all minor versions as major versions and each build version will be a minor version.
 
-## [Unreleased]
+## [0.5.2]
 ## Added
 - `JsonSerializable` to all entities - all entities can now be serialized to JSON
 - `markup` to `OrderItem`
 - `unitPrice` to `OrderItem`
 - `tip` to `Order`
 - `GetOrderInvoiceRequest` & `OrderInvoice`
+- `OrderDelivery` to `Order`
 
 ## [0.5.1]
 ### Fixed
@@ -66,7 +67,8 @@ For now, please use all minor versions as major versions and each build version 
 ## [< 0.3.1]
 ### History unavailable
 
-[Unreleased]: https://github.com/ordercloud/ordercloud-php/compare/0.5.1...HEAD
+[Unreleased]: https://github.com/ordercloud/ordercloud-php/compare/0.5.2...HEAD
+[0.5.2]: https://github.com/ordercloud/ordercloud-php/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/ordercloud/ordercloud-php/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/ordercloud/ordercloud-php/compare/0.4.2...0.5.0
 [0.4.2]: https://github.com/ordercloud/ordercloud-php/compare/0.4.1...0.4.2